requirements.txt
fastapi
uvicorn
python-dotenv
pyodbc
sqlglot